corrected delete logics finalised pdf implementation added bootstrap resources through webjars

+++ src/main/java/io/ankara/service/impl/TaxServiceBean.java
@@ -1,13 +1,18 @@
 package io.ankara.service.impl;
 
-import io.ankara.domain.Company;
-import io.ankara.domain.Tax;
+import io.ankara.domain.*;
+import io.ankara.repository.AppliedTaxRepository;
 import io.ankara.repository.TaxRepository;
+import io.ankara.service.EstimateService;
+import io.ankara.service.InvoiceService;
 import io.ankara.service.TaxService;
+import org.hibernate.validator.internal.engine.messageinterpolation.parser.ELState;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.inject.Inject;
+import java.util.Collection;
 import java.util.List;
 
 /**
@@ -17,10 +22,14 @@ import java.util.List;
  * @date 8/31/16 7:06 PM
  */
 @Service
+@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
 public class TaxServiceBean implements TaxService {
 
     @Inject
     private TaxRepository taxRepository;
+
+    @Inject
+    private AppliedTaxRepository appliedTaxRepository;
 
     @Transactional
     public boolean save(Tax tax) {
@@ -36,7 +45,12 @@ public class TaxServiceBean implements TaxService {
     @Override
     @Transactional
     public boolean delete(Tax tax) {
+        Collection<AppliedTax> appliedTaxes = appliedTaxRepository.findAllByTax(tax);
+        if(!appliedTaxes.isEmpty())
+            throw new IllegalStateException("There are invoices/estimates applied with this tax : "+tax);
+
         taxRepository.delete(tax);
+
         return true;
     }
 }
